Renamed mongo class file to mongo and fixed lint
